flood fill iterative approach

+++ src/canvas.h
@@ -14,10 +14,10 @@ public:
     bool drawing;
     SDL_Surface* surface;
     SDL_Surface* buffer;
-    std::array<float, 3> bodycolor;
+    std::array<Uint8, 3> bodycolor;
 
     // Constructor with default parameter values
-    Canvas(SDL_Renderer* renderer, int cx = 0, int cy = 0, int w = 720, int h = 480, float zoom = 1.0, bool drawing = false, const std::array<float, 3> bodycolor = { 255, 255, 255 });
+    Canvas(SDL_Renderer* renderer, int cx = 0, int cy = 0, int w = 720, int h = 480, float zoom = 1.0, bool drawing = false, const std::array<Uint8, 3> bodycolor = { 255, 255, 255 });
 
     // Member functions
     void pan(int x_off, int y_off);
