+++ src/line.h
@@ -8,8 +8,9 @@ class Line : public Tool {
 public:
     Line(Canvas& canvas);
 
-    void set_start(int x, int y, std::array<float, 3>& color) override;
-    void hold(int x, int y, std::array<float, 3>& color) override;
+    void set_start(int x, int y, std::array<Uint8, 3>& color) override;
+    void hold(int x, int y, std::array<Uint8, 3>& color) override;
 };
 
 #endif
+
